+++ piloteI2C1.h
@@ -6,6 +6,6 @@
 // Function prototypes for I2C operations
 int piloteI2C1_configureLAdresse(int fdPortI2C, int adresse);
 int piloteI2C1_ecritDesOctets(int *message, int taille);
-int piloteI2C1_litDesOctets(int *commande, int tailleCommande, uint8_t *donnees, int tailleDonnees);
+int piloteI2C1_litDesOctets(int *commande, int tailleCommande, int *donnees, int tailleDonnees);
 
 #endif // PILOTE_I2C1_H
